Feature sale feature

+++ database/migrations/2022_09_12_100029_create_purchase_items_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->unsignedBigInteger("inv_item_id");
             $table->unsignedBigInteger("conf_unit_id");
             $table->unsignedBigInteger("purchase_id");
-            $table->decimal("quantity");
-            $table->decimal("unit_price");
+            $table->float("quantity");
+            $table->decimal("unit_price",17,2);
             $table->softDeletes();
             $table->timestamps();
         });
